net: xilinx: axiethernet: Reduce the width of variable to handle overflow
Reduce the width of tx_bd_num variable to resolve the overflow
failure at dma_free_coherent and dma_alloc_coherent in
axienet_bd_free and dma_txq_init functions.

Addresses-Coverity: Event overflow_before_widen.
Acked-by: Harini Katakam <harini.katakam@amd.com>

+++ drivers/net/ethernet/xilinx/xilinx_axienet.h
@@ -767,7 +767,7 @@ struct axienet_local {
 	u32 options;
 	u32 features;
 
-	u32 tx_bd_num;
+	u16 tx_bd_num;
 	u32 rx_bd_num;
 
 	u32 max_frm_size;
